refactor : 이미지 파일 정리 및 API Index 정리
