+++ src/utils/handleUploadImage.js
@@ -1,4 +1,4 @@
-import { uploadImageTest } from "../api/imageTest";
+import { uploadImageChangeAPI } from "../api";
 import BASE_URL from "./BASE_URL";
 
 export const handleUploadImageAPI = async ({ files, inputFileElement }) => {
@@ -12,7 +12,7 @@ export const handleUploadImageAPI = async ({ files, inputFileElement }) => {
       size += files[i].size;
       const formData = new FormData();
       formData.append("image", files[i]);
-      resArray.push(uploadImageTest(formData).then((res) => BASE_URL + res.data.filename));
+      resArray.push(uploadImageChangeAPI(formData).then((res) => BASE_URL + res.data.filename));
     }
 
     if (size >= 10000000) {
